# Conflicts:
#	public_html/administrator/components/com_easydiscuss/themes/default/settings/antispam/general.php

+++ public_html/administrator/components/com_acym/classes/condition.php
@@ -9,7 +9,7 @@ class acymconditionClass extends acymClass
 
     public function getOneByStepId($stepId)
     {
-        $query = "SELECT * FROM #__acym_condition WHERE step_id = ".intval($stepId);
+        $query = 'SELECT * FROM #__acym_condition WHERE step_id = '.intval($stepId);
 
         return acym_loadObject($query);
     }
@@ -27,7 +27,7 @@ class acymconditionClass extends acymClass
 
         $actionClass = acym_get('class.action');
         $actionsIds = $actionClass->getAllActionsIdByConditionsId($elements);
-        $actionClass->delete($actionsIds);
+        if (!empty($actionsIds)) $actionClass->delete($actionsIds);
 
         return parent::delete($elements);
     }
